new test and common fixtures
