+++ services/web/server/tests/unit/with_dbs/02/test_projects_ports_handlers.py
@@ -4,6 +4,7 @@
 # pylint: disable=unused-variable
 
 from copy import deepcopy
+from datetime import datetime
 from typing import Any
 
 import pytest
@@ -13,13 +14,13 @@ from aiohttp.web_exceptions import HTTPOk
 from models_library.api_schemas_long_running_tasks.tasks import TaskStatus
 from models_library.api_schemas_webserver.projects import ProjectGet
 from models_library.generics import Envelope
-from models_library.projects_nodes import Node, NodeID
 from pydantic import parse_obj_as
 from pytest_simcore.helpers.faker_webserver import (
     PROJECTS_METADATA_PORTS_RESPONSE_BODY_DATA,
 )
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import UserInfoDict
+from pytest_simcore.helpers.utils_webserver_unit_with_db import MockedStorageSubsystem
 from servicelib.aiohttp.long_running_tasks.client import long_running_task_request
 from simcore_service_webserver.db.models import UserRole
 from simcore_service_webserver.projects.models import ProjectDict
@@ -27,135 +28,6 @@ from tenacity import TryAgain, retry
 from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 from yarl import URL
-
-
-@pytest.fixture
-def workbench_db_column() -> dict[str, Any]:
-    return {
-        "13220a1d-a569-49de-b375-904301af9295": {
-            "key": "simcore/services/comp/itis/sleeper",
-            "version": "2.1.4",
-            "label": "sleeper",
-            "inputs": {
-                "input_2": {
-                    "nodeUuid": "38a0d401-af4b-4ea7-ab4c-5005c712a546",
-                    "output": "out_1",
-                },
-                "input_3": False,
-                "input_4": 0,
-            },
-            "inputsUnits": {},
-            "inputNodes": ["38a0d401-af4b-4ea7-ab4c-5005c712a546"],
-            "parent": None,
-            "thumbnail": "",
-        },
-        "38a0d401-af4b-4ea7-ab4c-5005c712a546": {
-            "key": "simcore/services/frontend/parameter/integer",
-            "version": "1.0.0",
-            "label": "X",
-            "inputs": {},
-            "inputsUnits": {},
-            "inputNodes": [],
-            "parent": None,
-            "thumbnail": "",
-            "outputs": {"out_1": 43},
-            "runHash": None,
-        },
-        "08d15a6c-ae7b-4ea1-938e-4ce81a360ffa": {
-            "key": "simcore/services/comp/itis/sleeper",
-            "version": "2.1.4",
-            "label": "sleeper_2",
-            "inputs": {
-                "input_2": 2,
-                "input_3": {
-                    "nodeUuid": "7bf0741f-bae4-410b-b662-fc34b47c27c9",
-                    "output": "out_1",
-                },
-                "input_4": {
-                    "nodeUuid": "fc48252a-9dbb-4e07-bf9a-7af65a18f612",
-                    "output": "out_1",
-                },
-            },
-            "inputsUnits": {},
-            "inputNodes": [
-                "fc48252a-9dbb-4e07-bf9a-7af65a18f612",
-                "7bf0741f-bae4-410b-b662-fc34b47c27c9",
-            ],
-            "parent": None,
-            "thumbnail": "",
-            "state": {"currentStatus": "SUCCESS"},
-            "progress": 100,
-            "outputs": {
-                "output_1": {
-                    "store": 0,
-                    "path": "e08316a8-5afc-11ed-bab7-02420a00002b/08d15a6c-ae7b-4ea1-938e-4ce81a360ffa/single_number.txt",
-                    "eTag": "1679091c5a880faf6fb5e6087eb1b2dc",
-                },
-                "output_2": 6,
-            },
-            "runHash": "5d55ebe569aa0abeb5287104dc5989eabc755f160c9a5c9a1cc783fe1e058b66",
-        },
-        "fc48252a-9dbb-4e07-bf9a-7af65a18f612": {
-            "key": "simcore/services/frontend/parameter/integer",
-            "version": "1.0.0",
-            "label": "Z",
-            "inputs": {},
-            "inputsUnits": {},
-            "inputNodes": [],
-            "parent": None,
-            "thumbnail": "",
-            "outputs": {"out_1": 1},
-            "runHash": None,
-        },
-        "7bf0741f-bae4-410b-b662-fc34b47c27c9": {
-            "key": "simcore/services/frontend/parameter/boolean",
-            "version": "1.0.0",
-            "label": "on",
-            "inputs": {},
-            "inputsUnits": {},
-            "inputNodes": [],
-            "parent": None,
-            "thumbnail": "",
-            "outputs": {"out_1": False},
-            "runHash": None,
-        },
-        "09fd512e-0768-44ca-81fa-0cecab74ec1a": {
-            "key": "simcore/services/frontend/iterator-consumer/probe/integer",
-            "version": "1.0.0",
-            "label": "Random sleep interval_2",
-            "inputs": {
-                "in_1": {
-                    "nodeUuid": "13220a1d-a569-49de-b375-904301af9295",
-                    "output": "output_2",
-                }
-            },
-            "inputsUnits": {},
-            "inputNodes": ["13220a1d-a569-49de-b375-904301af9295"],
-            "parent": None,
-            "thumbnail": "",
-        },
-        "76f607b4-8761-4f96-824d-cab670bc45f5": {
-            "key": "simcore/services/frontend/iterator-consumer/probe/integer",
-            "version": "1.0.0",
-            "label": "Random sleep interval",
-            "inputs": {
-                "in_1": {
-                    "nodeUuid": "08d15a6c-ae7b-4ea1-938e-4ce81a360ffa",
-                    "output": "output_2",
-                }
-            },
-            "inputsUnits": {},
-            "inputNodes": ["08d15a6c-ae7b-4ea1-938e-4ce81a360ffa"],
-            "parent": None,
-            "thumbnail": "",
-        },
-    }
-
-
-@pytest.fixture
-def workbench(workbench_db_column: dict[str, Any]) -> dict[NodeID, Node]:
-    # convert to  model
-    return parse_obj_as(dict[NodeID, Node], workbench_db_column)
 
 
 @pytest.fixture
@@ -168,7 +40,7 @@ def fake_project(
     return project
 
 
-@pytest.mark.acceptance_test
+@pytest.mark.acceptance_test()
 @pytest.mark.parametrize(
     "user_role,expected",
     [
@@ -364,7 +236,7 @@ async def _wait_until_project_cloned_or_timeout(
     return task_result
 
 
-@pytest.mark.testit()
+@pytest.mark.xfail(reason="Under dev")
 @pytest.mark.parametrize(
     "user_role",
     [UserRole.USER],
@@ -373,7 +245,10 @@ async def test_clone_project_and_set_inputs(
     client: TestClient,
     logged_user: UserInfoDict,
     user_project: ProjectDict,
+    # mocks backend
+    storage_subsystem_mock: MockedStorageSubsystem,
     mock_catalog_service_api_responses: None,
+    project_db_cleaner: None,
 ):
     assert client.app
 
@@ -384,8 +259,9 @@ async def test_clone_project_and_set_inputs(
     assert f"/v0/projects/{parent_project_id}:clone" == url.path
 
     data = None
+
     async for long_running_task in long_running_task_request(
-        client.session, url, client_timeout=30
+        client.session, url=client.make_url(url.path), json=None, client_timeout=30
     ):
         print(f"{long_running_task.progress=}")
         if long_running_task.done():
@@ -396,7 +272,9 @@ async def test_clone_project_and_set_inputs(
 
     assert parent_project_id != cloned_project.uuid
     assert user_project["description"] == cloned_project.description
-    assert user_project["creation_date"] < cloned_project.creation_date
+    assert parse_obj_as(datetime, user_project["creationDate"]) < parse_obj_as(
+        datetime, cloned_project.creation_date
+    )
 
     # - set_inputs project_clone_id ----------------------------------------------
     job_inputs_values = {"X": 42}  # like JobInputs.values
